started deprecation and additional clean up
