+++ .github/workflows/pylint.yml
@@ -20,4 +20,4 @@ jobs:
           pip install pylint
       - name: Analysing the code with pylint
         run: |
-          pylint $(git ls-files '*.py') --fail-under=8 --disable=R0912,R0913,R0914,R0915,E0401,E0402,R0801,C0103
+          pylint $(git ls-files '*.py') --fail-under=8 --ignore=virga/stage_deprecated.py
